halfDay feature added.

+++ internshipdemo/src/main/java/com/kafein/internshipdemo/service/EmployeeService.java
@@ -18,7 +18,7 @@ public class EmployeeService implements IEmployeeService {
     private EmployeeRepository employeeRepository;
 
    @Value("${employee.break.duration}")
-   Integer breakDuration;
+   Double breakDuration;
 
     @Autowired
     public EmployeeService(EmployeeRepository employeeRepository) {
